Consolidate some obo tests (no code change)
